+++ src/commands/roll.ts
@@ -1,6 +1,16 @@
 import BaseCommand, { ExecContext } from './base';
 import { getPrefix } from '../utils/misc';
+import { reactFail } from '../utils/discord';
 
+/**
+ * A function that returns a random floating point
+ * number from 0 to 1.
+ */
+type RandomFunc = () => number;
+
+/**
+ * Various RNG commands.
+ */
 export default class Roll implements BaseCommand {
     public async exec(ctx: ExecContext) {
         const { msg, args } = ctx;
@@ -15,16 +25,21 @@ export default class Roll implements BaseCommand {
 
         if (!args.length) {
             // Return a random number from 0-100.
-            this.sendMessage(ctx, this.intRange().toString());
+            this.sendMessage(ctx, Roll.intRange().toString());
             return;
         }
 
-        if (args.length === 1 && !Number.isNaN(firstArgInt)) {
-            // Get a number from 0 to the provided number
-            this.sendMessage(
-                ctx,
-                this.intRange(0, firstArgInt + 1).toString(),
-            );
+        if (args.length === 1) {
+            if (!Number.isNaN(firstArgInt)) {
+                // Get a number from 0 to the provided number
+                this.sendMessage(
+                    ctx,
+                    Roll.intRange(0, firstArgInt + 1).toString(),
+                );
+                return;
+            }
+
+            reactFail(msg, 'if you are providing just 1 argument, it has to be a number!');
             return;
         }
 
@@ -33,7 +48,7 @@ export default class Roll implements BaseCommand {
             // Argument is a number prefixed with d, e.g. d6 - do a dice roll.
             this.sendMessage(
                 ctx,
-                this.intRange(1, firstArgSliceInt + 1).toString(),
+                Roll.intRange(1, firstArgSliceInt + 1).toString(),
             );
             return;
         }
@@ -42,13 +57,13 @@ export default class Roll implements BaseCommand {
             // Get a number between the provided numbers
             this.sendMessage(
                 ctx,
-                this.intRange(firstArgInt, secondArgInt).toString(),
+                Roll.intRange(firstArgInt, secondArgInt).toString(),
             );
             return;
         }
 
         // Random choice between given options
-        this.sendMessage(ctx, args[this.intRange(0, args.length)]);
+        this.sendMessage(ctx, args[Roll.intRange(0, args.length)]);
     }
 
     private sendMessage(ctx: ExecContext, message: string) {
@@ -60,8 +75,15 @@ export default class Roll implements BaseCommand {
      * Upper bound exclusive.
      * 0 to 100 by default.
      */
-    private intRange(from: number = 0, to: number = 101): number {
-        return Math.floor(Math.random() * (to - from)) + from;
+    public static intRange(
+        from: number = 0,
+        to: number = 101,
+        randomFunc: RandomFunc = Math.random,
+    ): number {
+        if (from > to) {
+            return Math.floor(randomFunc() * (from - to)) + to;
+        }
+        return Math.floor(randomFunc() * (to - from)) + from;
     }
 
     public async update(): Promise<void> {}